Added event buses for Domain and Pojo accross Sticky, Consumable and Queued
Domain = real objects saved in bus that may be set in a service and
consumed in the UI.
Pojo = intent serialised beans that are sent from the service and
rehydrated in the event bus in the UI tier which may be consumed by the
UI (lower latency).

Sticky = once the event is set, it is retained by the event bus until a
new event is sent.
Consumable = once an event is used by the consumer, the event is
removed.  Subsequent calls will return null.
Queued = an in order FIFO queue of events that can be picked up in the
UI tier.